test: 테스트를 구매, 당첨번호 입력, 모달에 따라 분리
